+++ src/browser/index.ts
@@ -1,4 +1,3 @@
-export { json, xml } from "../core/parsers";
 export { BrowserClient } from "./client";
 export { csv, tsv } from "./parsers";
 export { fileReader } from "./readers";