+++ src/browser/parsers.ts
@@ -1,4 +1,4 @@
+import { createDelimitedParsers } from "@openaq/transform/core";
 import { parse } from "csv-parse/browser/esm/sync";
-import { createDelimitedParsers } from "../core/parsers";
 
 export const { csv, tsv } = createDelimitedParsers(parse);