+++ src/node/index.ts
@@ -1,4 +1,3 @@
-export { json, xml } from "../core/parsers";
 export { NodeClient } from "./client";
 export { csv, tsv } from "./parsers";
 export { fileSystemReader } from "./readers";